// monopoly.client/src/app/models/game-objects.model.ts
import { StreetGameObject } from "./street-object.model";

export interface IGameObjects {
    greyStreet: StreetGameObject[];
    pinkStreet: StreetGameObject[];
    yellowStreet: StreetGameObject[];
    greenStreet: StreetGameObject[];
    blueStreet: StreetGameObject[];
    cornStreet: StreetGameObject[];
    orangeStreet: StreetGameObject[];
    redStreet: StreetGameObject[];
    gameObjectIds: string[];
}